all the tester for mandatory are done

+++ sources/libft.h
@@ -69,8 +69,8 @@ char	*ft_convert_base(char *nbr, char *base_from, char *base_to);
 int		ft_atoi_base(char *str, char *base);
 int		checkbase(char *base);
 char	*ft_llutoa(unsigned long long num);
-long	ft_len_nbr(long nbr, char *base);
-void	ft_putnbr_base(long int nbr, char *base);
+long	ft_len_nbr(long long nbr, char *base);
+void	ft_putnbr_base(unsigned long long nbr, char *base);
 char	*ft_utoa(int n);
 
 #endif
